+++ spec/fixtures/sleep.rb
@@ -1,0 +1,1 @@
+sleep(ARGV.first.to_i || 0)